+++ app/assets/javascripts/Components/Helpers/range_selector.js
@@ -54,12 +54,72 @@ function getLeafNodes(root, _nodes) {
   return _nodes;
 }
 
-export function markupTextInRange(range, colour) {
+export function pathToNode(node) {
+  if (node.id) {
+    return `//*[@id="${node.id}"]`;
+  } else if (node.parentNode) {
+    if (node.nodeType === Node.TEXT_NODE) {
+      const index = [...node.parentNode.childNodes]
+        .filter(n => n.nodeType === Node.TEXT_NODE)
+        .indexOf(node);
+      return `${pathToNode(node.parentNode)}/text()[${index + 1}]`;
+    } else {
+      const index = [...node.parentNode.childNodes]
+        .filter(n => n.tagName === node.tagName)
+        .indexOf(node);
+      return `${pathToNode(node.parentNode)}/${node.tagName}[${index + 1}]`;
+    }
+  } else {
+    return "";
+  }
+}
+
+function addMouseOverToNode(node, content) {
+  const content_container = node.ownerDocument.createElement("div");
+  Object.assign(content_container.style, {
+    display: "none",
+    // copy of the css properties for .annotation_text_display elements
+    background: document.documentElement.style.getPropertyValue("--background_main"),
+    border: `1px solid ${document.documentElement.style.getPropertyValue("--sharp_line")}`,
+    borderRadius: getComputedStyle(document.documentElement).getPropertyValue("--radius"),
+    boxShadow: `4px 4px 2px ${document.documentElement.style.getPropertyValue("--primary_two")}`,
+    maxWidth: "400px",
+    padding: "0.25em 1em",
+    position: "absolute",
+    width: "auto",
+    wordWrap: "break-word",
+    zIndex: "100000",
+  });
+  content_container.innerHTML = safe_marked(content);
+  content_container.className = "markus-annotation-content";
+  node.ownerDocument.body.appendChild(content_container);
+  // TODO: apply MathJax typesetting to the content_container node
+  //       MathJax.Hub.Queue(["Typeset", MathJax.Hub, content_container]); // <- this works but mathjax css isn't applied
+  //                                                                       //    because iframe has its own css context
+  node.addEventListener("mouseenter", e => {
+    let offset_height = 0;
+    for (let elem of node.ownerDocument.getElementsByClassName("markus-annotation-content")) {
+      if (getComputedStyle(elem).display !== "none") {
+        offset_height += elem.offsetHeight;
+      }
+    }
+    content_container.style.left = `${e.pageX}px`;
+    content_container.style.top = `${e.pageY + offset_height}px`;
+    content_container.style.display = "";
+  });
+  node.addEventListener("mouseleave", () => {
+    content_container.style.display = "none";
+  });
+}
+
+export function markupTextInRange(range, colour, content) {
   if (range.startContainer === range.endContainer) {
     const old_node = range.startContainer;
     const parent = old_node.parentNode;
-    if (old_node.nodeType === 3) {
-      const new_node = document.createElement("span");
+    let new_node;
+    if (old_node.nodeType === Node.TEXT_NODE) {
+      new_node = document.createElement("span");
+      new_node.className = "markus-annotation";
       new_node.style.backgroundColor = colour;
       const unmarked1 = document.createTextNode(old_node.nodeValue.substring(0, range.startOffset));
       const marked = document.createTextNode(
@@ -71,17 +131,21 @@ export function markupTextInRange(range, colour) {
       parent.insertBefore(new_node, unmarked1.nextSibling);
       parent.insertBefore(unmarked2, new_node.nextSibling);
     } else if (old_node.nodeName === "img" || old_node.childNodes.length) {
-      const new_node = document.createElement("div");
+      new_node = document.createElement("div");
+      new_node.className = "markus-annotation";
       new_node.style.border = `5px solid ${colour}`;
       new_node.appendChild(old_node.cloneNode(true));
       parent.replaceChild(new_node, old_node);
     }
+    addMouseOverToNode(new_node, content);
   } else {
     getAllNodesInRange(range).forEach(node => {
       getLeafNodes(node).forEach(old_node => {
         const parent = old_node.parentNode;
-        if (old_node.nodeType === 3) {
-          const new_node = document.createElement("span");
+        let new_node;
+        if (old_node.nodeType === Node.TEXT_NODE) {
+          new_node = document.createElement("span");
+          new_node.className = "markus-annotation";
           new_node.style.backgroundColor = colour;
           if (old_node === range.startContainer) {
             const unmarked = document.createTextNode(
@@ -104,11 +168,13 @@ export function markupTextInRange(range, colour) {
             parent.replaceChild(new_node, old_node);
           }
         } else if (old_node.nodeName === "img" || old_node.childNodes.length) {
-          const new_node = document.createElement("div");
+          new_node = document.createElement("div");
+          new_node.className = "markus-annotation";
           new_node.style.border = `5px solid ${colour}`;
           new_node.appendChild(old_node.cloneNode(true));
           parent.replaceChild(new_node, old_node);
         }
+        addMouseOverToNode(new_node, content);
       });
     });
   }