+++ app/policies/html_annotation_policy.rb
@@ -1,0 +1,3 @@
+# Html annotation policy class
+class HtmlAnnotationPolicy < AnnotationPolicy
+end