+++ db/migrate/20220105204341_add_columns_for_html_annotations.rb
@@ -1,0 +1,8 @@
+class AddColumnsForHtmlAnnotations < ActiveRecord::Migration[6.1]
+  def change
+    add_column :annotations, :start_node, :string
+    add_column :annotations, :end_node, :string
+    add_column :annotations, :start_offset, :integer
+    add_column :annotations, :end_offset, :integer
+  end
+end