+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## [unreleased]
 - Remove unmaintained locales (#5727)
+- Add ability to annotate notebook (jupyter and Rmd) submissions (#5749)
 
 ## [v2.0.4]
 - Fix bug where "Create Note" button was displaying when a course had no noteables (#5745)